+++ src/components/auth/loginForm/Step1Form.tsx
@@ -1,6 +1,6 @@
 "use client";
-import Button from "@/components/common/Button/Button";
-import TextInput from "@/components/common/TextInput";
+import Button from "@/components/common/button/Button";
+import TextInput from "@/components/common/textInput/TextInput";
 import { Cookie } from "@/interfaces/login";
 import { shemaEmail } from "@/schema/login";
 import { yupResolver } from "@hookform/resolvers/yup";
